Revert "use hpcopy instead of libdmg-hfsplus"
This reverts commit 5894e50d88f2afbd3b41b5682defd8590c7089c4.

+++ distribution/macos/create_macos_pr_build_ava.sh
@@ -114,13 +114,31 @@ echo "Staging directory for packaging"
 DMG_FOLDER="$OUTPUT_DIRECTORY/dmg"
 mkdir "$DMG_FOLDER"
 
+echo ""
+echo "Checking exec perms..."
+ls -l "$UNIVERSAL_APP_BUNDLE/Contents/MacOS/Ryujinx"
+echo ""
+
 tar -xzf "$BASE_DIRECTORY/distribution/macos/DMG_ASSETS/DMG_Structure.tar.gz" -C "$DMG_FOLDER" --strip-components=1
 cp -R "$UNIVERSAL_APP_BUNDLE" "$DMG_FOLDER/Ryujinx.app"
 
 # Set permissions explicitly, based on how macOS expects them.
 find "$DMG_FOLDER" -type d -exec chmod 0755 {} +
 find "$DMG_FOLDER" -type f -exec chmod 0644 {} +
-chmod 0755 "$DMG_FOLDER/Ryujinx.app/Contents/MacOS/Ryujinx"
+
+find "$DMG_FOLDER" -mindepth 1 -type f | while IFS= read -r src;
+do
+    dst="/${src#"$DMG_FOLDER"/}"
+    if file "$src" | grep -Fqi "Mach-O"; 
+    then
+        chmod 0755 "$src"
+    fi
+done
+
+echo ""
+echo "Checking exec perms..."
+ls -l "$DMG_FOLDER/Ryujinx.app/Contents/MacOS/Ryujinx"
+echo ""
 
 # Now sign it.
 echo ""
@@ -142,6 +160,11 @@ else
     echo "Using codesign to verify signature"
     spctl -a -vv "$DMG_FOLDER/Ryujinx.app"
 fi
+
+echo ""
+echo "Checking exec perms..."
+ls -l "$DMG_FOLDER/Ryujinx.app/Contents/MacOS/Ryujinx"
+echo ""
 
 # Create archive for legacy releases.
 echo ""
@@ -167,27 +190,40 @@ TOTAL_SIZE=$((STAGING_SIZE + PADDING))
 dd if=/dev/zero of="$UNCOMPRESSED_DMG" bs=1 count=0 seek="$TOTAL_SIZE" status=none
 mkfs.hfsplus -v "Ryujinx" "$UNCOMPRESSED_DMG"
 
-hpmount "$UNCOMPRESSED_DMG"
-# Make all the folders first, because hpcopy won't make non-existent directories.
+# Make all the folders first, because libdmg-hfsplus won't make non-existent directories.
 find "$DMG_FOLDER" -mindepth 1 -type d | sort | while IFS= read -r src;
 do
     dst="/${src#"$DMG_FOLDER"/}"
-    hpmkdir "$dst"
+    dmg-hfsplus "$UNCOMPRESSED_DMG" mkdir "$dst"
 done
-hpcopy "$DMG_FOLDER/" /
-hpumount
+# Copy the files over.
+find "$DMG_FOLDER" -mindepth 1 -type f | while IFS= read -r src;
+do
+    dst="/${src#"$DMG_FOLDER"/}"
+    dmg-hfsplus "$UNCOMPRESSED_DMG" add "$src" "$dst"
+done
+
+echo ""
+echo "Checking exec perms..."
+hfsplus "$UNCOMPRESSED_DMG" extract /Ryujinx.app/Content/MacOS/Ryujinx "$BASE_DIRECTORY/Ryujinx"
+ls -l "$BASE_DIRECTORY/Ryujinx"
+rm -f "$BASE_DIRECTORY/Ryujinx"
+echo ""
 
 # Copy the symlink into a folder, then copy the folder over with symlink permissions.
 # It doesn't like files that much.
-TEMP="$OUTPUT_DIRECTORY/temp"
-mkdir -p "$TEMP"
-cp -P "$DMG_FOLDER/Applications" "$TEMP"
-hfsplus "$UNCOMPRESSED_DMG" -s clone_link addall "$TEMP" /
-rm -rf "$TEMP"
+if [[ -L "$DMG_FOLDER/Applications" ]]; 
+then
+    TEMP="$OUTPUT_DIRECTORY/temp"
+    mkdir -p "$TEMP"
+    cp -P "$DMG_FOLDER/Applications" "$TEMP"
+    dmg-hfsplus "$UNCOMPRESSED_DMG" -s clone_link addall "$TEMP" /
+    rm -rf "$TEMP"
+fi
 
 # https://developer.apple.com/library/archive/technotes/tn/tn1150.html
 # kHasCustomIcon
-hfsplus "$UNCOMPRESSED_DMG" attr / C
+dmg-hfsplus "$UNCOMPRESSED_DMG" attr / C
 dmg dmg -c lzma "$UNCOMPRESSED_DMG" "$COMPRESSED_DMG"
 
 rm -f "$UNCOMPRESSED_DMG"
